requirements.txt
streamlit
yfinance
pandas
pyyaml 